Run an SEIR model with Erlang passage times

Wrapper function for an SEIR compartmental model with Erlang passage times based on Getz and Dougherty (2017) J. Biological Dynamics, and developed to model the West African ebola virus disease outbreak of 2014 .

Usage

.epidemic_ebola_cpp(
  initial_state,
  population_size,
  beta,
  shape_E,
  rate_E,
  shape_I,
  rate_I,
  time_end
)

Arguments

population_size

A single integer for the population size.

beta

The transmission rate \(\beta\).

shape_E

A single integer for the shape parameter of the Erlang distribution of passage times through the exposed compartment.

rate_E

A single double for the rate parameter of the Erlang distribution of passage times through the exposed compartment.

shape_I

A single integer for the shape parameter of the Erlang distribution of passage times through the infectious compartment.

rate_I

A single double for the rate parameter of the Erlang distribution of passage times through the infectious compartment.

time_end

A single integer for the maximum simulation time; this is assumed to be in days.

initial_conditions

An integer vector of the initial conditions corresponding to an SEIR model.

Value

A list with two elements, x, and integer matrix with as many rows as the number of timesteps (given by time_end), and four columns, one for each compartment, susceptible, exposed, infectious, recovered, in that order ; and times, a vector of the simulation times, taken to be days. This output is intended to be passed to output_to_df() to be converted into a data.frame for further analysis.