Model an SIRV epidemic

Function for a deterministic susceptible-infectious-recovered model with an optional vaccination component. Allows heterogeneity in social contact patterns, the demography distribution, and key epidemiological parameters: the reproductive number \(R_0\), and the infectious period \(1/\gamma\). Also allows for group-specific initial proportions in each model compartment, as well as group-specific vaccination start dates and vaccination rates.

Usage

epi_demic(
  population,
  model = "default",
  R0 = 1.5,
  infectious_period = 7,
  intervention,
  nu = 0.01,
  t_vax_begin = t_max * 0.6,
  t_vax_end = t_max * 0.8,
  t_max = 200,
  t_increment = 1
)

Arguments

population

An object of the population class, which holds a population contact matrix, a demography vector, and the initial conditions of each demographic group. See population().

model

Epidemic model function.

R0

The reproductive number of the infection. May be a single number or a vector of the same length as the number of demographic groups.

infectious_period

The mean infections period. May be a single number or a vector of the same length as the number of demographic groups.

intervention

An object of the intervention class that specifies the start and end of the intervention, and the effect of the intervention on each demographic group. See intervention().

nu

The vaccination rate \(\nu\). May be a single number or a vector of the same length as the number of demographic groups.

t_vax_begin

The time at which vaccination begins. Must be a number, less than t_max. Defaults to 60% of the maximum time.

t_vax_end

The time at which vaccination ends. Defaults to 80% of the maximum time.

t_max

The maximum number of timesteps over which to run the model.

t_increment

The size of the time increment.

Value

A data.frame with the columns "time", "compartment", "age_group", "value". The compartments are "susceptible", "infected", "recovered", and "vaccinated".

Examples

# create a population
population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0),
    nrow = 1, ncol = 4
  )
)

# set up intervention
new_intervention <- intervention(
  name = "close schools",
  time_begin = 50,
  time_end = 80,
  contact_reduction = c(0.2)
)

# run epidemic simulation
epi_demic(
  population = population,
  R0 = 1.5, infectious_period = 7,
  intervention = new_intervention,
  nu = 0.01,
  t_vax_begin = 100,
  t_vax_end = 200,
  t_max = 200,
  t_increment = 1
)
#>      time compartment age_group        value
#> 11      0 susceptible         1 6.699330e+07
#> 12      1 susceptible         1 6.699181e+07
#> 13      2 susceptible         1 6.699021e+07
#> 14      3 susceptible         1 6.698850e+07
#> 15      4 susceptible         1 6.698665e+07
#> 16      5 susceptible         1 6.698467e+07
#> 17      6 susceptible         1 6.698255e+07
#> 18      7 susceptible         1 6.698026e+07
#> 19      8 susceptible         1 6.697781e+07
#> 110     9 susceptible         1 6.697518e+07
#> 111    10 susceptible         1 6.697235e+07
#> 112    11 susceptible         1 6.696931e+07
#> 113    12 susceptible         1 6.696605e+07
#> 114    13 susceptible         1 6.696255e+07
#> 115    14 susceptible         1 6.695879e+07
#> 116    15 susceptible         1 6.695475e+07
#> 117    16 susceptible         1 6.695041e+07
#> 118    17 susceptible         1 6.694576e+07
#> 119    18 susceptible         1 6.694076e+07
#> 120    19 susceptible         1 6.693539e+07
#> 121    20 susceptible         1 6.692962e+07
#> 122    21 susceptible         1 6.692343e+07
#> 123    22 susceptible         1 6.691679e+07
#> 124    23 susceptible         1 6.690965e+07
#> 125    24 susceptible         1 6.690199e+07
#> 126    25 susceptible         1 6.689376e+07
#> 127    26 susceptible         1 6.688494e+07
#> 128    27 susceptible         1 6.687546e+07
#> 129    28 susceptible         1 6.686528e+07
#> 130    29 susceptible         1 6.685436e+07
#> 131    30 susceptible         1 6.684264e+07
#> 132    31 susceptible         1 6.683006e+07
#> 133    32 susceptible         1 6.681655e+07
#> 134    33 susceptible         1 6.680206e+07
#> 135    34 susceptible         1 6.678651e+07
#> 136    35 susceptible         1 6.676982e+07
#> 137    36 susceptible         1 6.675192e+07
#> 138    37 susceptible         1 6.673271e+07
#> 139    38 susceptible         1 6.671210e+07
#> 140    39 susceptible         1 6.668999e+07
#> 141    40 susceptible         1 6.666628e+07
#> 142    41 susceptible         1 6.664085e+07
#> 143    42 susceptible         1 6.661357e+07
#> 144    43 susceptible         1 6.658433e+07
#> 145    44 susceptible         1 6.655298e+07
#> 146    45 susceptible         1 6.651937e+07
#> 147    46 susceptible         1 6.648335e+07
#> 148    47 susceptible         1 6.644474e+07
#> 149    48 susceptible         1 6.640338e+07
#> 150    49 susceptible         1 6.635906e+07
#> 151    50 susceptible         1 6.631160e+07
#> 152    51 susceptible         1 6.627179e+07
#> 153    52 susceptible         1 6.623093e+07
#> 154    53 susceptible         1 6.618900e+07
#> 155    54 susceptible         1 6.614596e+07
#> 156    55 susceptible         1 6.610181e+07
#> 157    56 susceptible         1 6.605651e+07
#> 158    57 susceptible         1 6.601004e+07
#> 159    58 susceptible         1 6.596238e+07
#> 160    59 susceptible         1 6.591351e+07
#> 161    60 susceptible         1 6.586339e+07
#> 162    61 susceptible         1 6.581201e+07
#> 163    62 susceptible         1 6.575935e+07
#> 164    63 susceptible         1 6.570537e+07
#> 165    64 susceptible         1 6.565006e+07
#> 166    65 susceptible         1 6.559339e+07
#> 167    66 susceptible         1 6.553534e+07
#> 168    67 susceptible         1 6.547589e+07
#> 169    68 susceptible         1 6.541500e+07
#> 170    69 susceptible         1 6.535266e+07
#> 171    70 susceptible         1 6.528885e+07
#> 172    71 susceptible         1 6.522354e+07
#> 173    72 susceptible         1 6.515671e+07
#> 174    73 susceptible         1 6.508833e+07
#> 175    74 susceptible         1 6.501840e+07
#> 176    75 susceptible         1 6.494687e+07
#> 177    76 susceptible         1 6.487374e+07
#> 178    77 susceptible         1 6.479899e+07
#> 179    78 susceptible         1 6.472259e+07
#> 180    79 susceptible         1 6.464452e+07
#> 181    80 susceptible         1 6.456478e+07
#> 182    81 susceptible         1 6.446085e+07
#> 183    82 susceptible         1 6.435031e+07
#> 184    83 susceptible         1 6.423280e+07
#> 185    84 susceptible         1 6.410794e+07
#> 186    85 susceptible         1 6.397534e+07
#> 187    86 susceptible         1 6.383459e+07
#> 188    87 susceptible         1 6.368528e+07
#> 189    88 susceptible         1 6.352699e+07
#> 190    89 susceptible         1 6.335929e+07
#> 191    90 susceptible         1 6.318174e+07
#> 192    91 susceptible         1 6.299391e+07
#> 193    92 susceptible         1 6.279535e+07
#> 194    93 susceptible         1 6.258562e+07
#> 195    94 susceptible         1 6.236429e+07
#> 196    95 susceptible         1 6.213092e+07
#> 197    96 susceptible         1 6.188510e+07
#> 198    97 susceptible         1 6.162642e+07
#> 199    98 susceptible         1 6.135449e+07
#> 1100   99 susceptible         1 6.106896e+07
#> 1101  100 susceptible         1 6.076949e+07
#> 1102  101 susceptible         1 5.985434e+07
#> 1103  102 susceptible         1 5.893779e+07
#> 1104  103 susceptible         1 5.802057e+07
#> 1105  104 susceptible         1 5.710345e+07
#> 1106  105 susceptible         1 5.618731e+07
#> 1107  106 susceptible         1 5.527306e+07
#> 1108  107 susceptible         1 5.436167e+07
#> 1109  108 susceptible         1 5.345415e+07
#> 1110  109 susceptible         1 5.255155e+07
#> 1111  110 susceptible         1 5.165492e+07
#> 1112  111 susceptible         1 5.076531e+07
#> 1113  112 susceptible         1 4.988378e+07
#> 1114  113 susceptible         1 4.901136e+07
#> 1115  114 susceptible         1 4.814905e+07
#> 1116  115 susceptible         1 4.729780e+07
#> 1117  116 susceptible         1 4.645853e+07
#> 1118  117 susceptible         1 4.563207e+07
#> 1119  118 susceptible         1 4.481921e+07
#> 1120  119 susceptible         1 4.402066e+07
#> 1121  120 susceptible         1 4.323705e+07
#> 1122  121 susceptible         1 4.246893e+07
#> 1123  122 susceptible         1 4.171677e+07
#> 1124  123 susceptible         1 4.098095e+07
#> 1125  124 susceptible         1 4.026179e+07
#> 1126  125 susceptible         1 3.955951e+07
#> 1127  126 susceptible         1 3.887427e+07
#> 1128  127 susceptible         1 3.820613e+07
#> 1129  128 susceptible         1 3.755510e+07
#> 1130  129 susceptible         1 3.692114e+07
#> 1131  130 susceptible         1 3.630411e+07
#> 1132  131 susceptible         1 3.570385e+07
#> 1133  132 susceptible         1 3.512014e+07
#> 1134  133 susceptible         1 3.455270e+07
#> 1135  134 susceptible         1 3.400124e+07
#> 1136  135 susceptible         1 3.346540e+07
#> 1137  136 susceptible         1 3.294483e+07
#> 1138  137 susceptible         1 3.243914e+07
#> 1139  138 susceptible         1 3.194790e+07
#> 1140  139 susceptible         1 3.147070e+07
#> 1141  140 susceptible         1 3.100710e+07
#> 1142  141 susceptible         1 3.055664e+07
#> 1143  142 susceptible         1 3.011889e+07
#> 1144  143 susceptible         1 2.969339e+07
#> 1145  144 susceptible         1 2.927968e+07
#> 1146  145 susceptible         1 2.887733e+07
#> 1147  146 susceptible         1 2.848588e+07
#> 1148  147 susceptible         1 2.810492e+07
#> 1149  148 susceptible         1 2.773400e+07
#> 1150  149 susceptible         1 2.737272e+07
#> 1151  150 susceptible         1 2.702066e+07
#> 1152  151 susceptible         1 2.667745e+07
#> 1153  152 susceptible         1 2.634270e+07
#> 1154  153 susceptible         1 2.601603e+07
#> 1155  154 susceptible         1 2.569710e+07
#> 1156  155 susceptible         1 2.538557e+07
#> 1157  156 susceptible         1 2.508112e+07
#> 1158  157 susceptible         1 2.478342e+07
#> 1159  158 susceptible         1 2.449218e+07
#> 1160  159 susceptible         1 2.420712e+07
#> 1161  160 susceptible         1 2.392797e+07
#> 1162  161 susceptible         1 2.365446e+07
#> 1163  162 susceptible         1 2.338635e+07
#> 1164  163 susceptible         1 2.312340e+07
#> 1165  164 susceptible         1 2.286541e+07
#> 1166  165 susceptible         1 2.261215e+07
#> 1167  166 susceptible         1 2.236343e+07
#> 1168  167 susceptible         1 2.211906e+07
#> 1169  168 susceptible         1 2.187886e+07
#> 1170  169 susceptible         1 2.164267e+07
#> 1171  170 susceptible         1 2.141032e+07
#> 1172  171 susceptible         1 2.118168e+07
#> 1173  172 susceptible         1 2.095659e+07
#> 1174  173 susceptible         1 2.073493e+07
#> 1175  174 susceptible         1 2.051657e+07
#> 1176  175 susceptible         1 2.030140e+07
#> 1177  176 susceptible         1 2.008930e+07
#> 1178  177 susceptible         1 1.988018e+07
#> 1179  178 susceptible         1 1.967392e+07
#> 1180  179 susceptible         1 1.947045e+07
#> 1181  180 susceptible         1 1.926967e+07
#> 1182  181 susceptible         1 1.907151e+07
#> 1183  182 susceptible         1 1.887588e+07
#> 1184  183 susceptible         1 1.868272e+07
#> 1185  184 susceptible         1 1.849196e+07
#> 1186  185 susceptible         1 1.830354e+07
#> 1187  186 susceptible         1 1.811738e+07
#> 1188  187 susceptible         1 1.793345e+07
#> 1189  188 susceptible         1 1.775168e+07
#> 1190  189 susceptible         1 1.757202e+07
#> 1191  190 susceptible         1 1.739443e+07
#> 1192  191 susceptible         1 1.721886e+07
#> 1193  192 susceptible         1 1.704528e+07
#> 1194  193 susceptible         1 1.687363e+07
#> 1195  194 susceptible         1 1.670388e+07
#> 1196  195 susceptible         1 1.653599e+07
#> 1197  196 susceptible         1 1.636994e+07
#> 1198  197 susceptible         1 1.620568e+07
#> 1199  198 susceptible         1 1.604319e+07
#> 1200  199 susceptible         1 1.588244e+07
#> 1201  200 susceptible         1 1.572339e+07
#> 21      0    infected         1 6.700000e+03
#> 22      1    infected         1 7.195907e+03
#> 23      2    infected         1 7.728480e+03
#> 24      3    infected         1 8.300425e+03
#> 25      4    infected         1 8.914647e+03
#> 26      5    infected         1 9.574261e+03
#> 27      6    infected         1 1.028261e+04
#> 28      7    infected         1 1.104330e+04
#> 29      8    infected         1 1.186016e+04
#> 210     9    infected         1 1.273735e+04
#> 211    10    infected         1 1.367930e+04
#> 212    11    infected         1 1.469076e+04
#> 213    12    infected         1 1.577686e+04
#> 214    13    infected         1 1.694306e+04
#> 215    14    infected         1 1.819527e+04
#> 216    15    infected         1 1.953977e+04
#> 217    16    infected         1 2.098334e+04
#> 218    17    infected         1 2.253324e+04
#> 219    18    infected         1 2.419724e+04
#> 220    19    infected         1 2.598369e+04
#> 221    20    infected         1 2.790154e+04
#> 222    21    infected         1 2.996038e+04
#> 223    22    infected         1 3.217047e+04
#> 224    23    infected         1 3.454283e+04
#> 225    24    infected         1 3.708926e+04
#> 226    25    infected         1 3.982240e+04
#> 227    26    infected         1 4.275579e+04
#> 228    27    infected         1 4.590390e+04
#> 229    28    infected         1 4.928226e+04
#> 230    29    infected         1 5.290748e+04
#> 231    30    infected         1 5.679731e+04
#> 232    31    infected         1 6.097076e+04
#> 233    32    infected         1 6.544814e+04
#> 234    33    infected         1 7.025117e+04
#> 235    34    infected         1 7.540307e+04
#> 236    35    infected         1 8.092860e+04
#> 237    36    infected         1 8.685425e+04
#> 238    37    infected         1 9.320825e+04
#> 239    38    infected         1 1.000207e+05
#> 240    39    infected         1 1.073238e+05
#> 241    40    infected         1 1.151516e+05
#> 242    41    infected         1 1.235407e+05
#> 243    42    infected         1 1.325298e+05
#> 244    43    infected         1 1.421601e+05
#> 245    44    infected         1 1.524755e+05
#> 246    45    infected         1 1.635223e+05
#> 247    46    infected         1 1.753500e+05
#> 248    47    infected         1 1.880107e+05
#> 249    48    infected         1 2.015598e+05
#> 250    49    infected         1 2.160558e+05
#> 251    50    infected         1 2.315602e+05
#> 252    51    infected         1 2.378409e+05
#> 253    52    infected         1 2.442661e+05
#> 254    53    infected         1 2.508382e+05
#> 255    54    infected         1 2.575593e+05
#> 256    55    infected         1 2.644309e+05
#> 257    56    infected         1 2.714547e+05
#> 258    57    infected         1 2.786325e+05
#> 259    58    infected         1 2.859655e+05
#> 260    59    infected         1 2.934554e+05
#> 261    60    infected         1 3.011033e+05
#> 262    61    infected         1 3.089103e+05
#> 263    62    infected         1 3.168777e+05
#> 264    63    infected         1 3.250061e+05
#> 265    64    infected         1 3.332965e+05
#> 266    65    infected         1 3.417494e+05
#> 267    66    infected         1 3.503653e+05
#> 268    67    infected         1 3.591443e+05
#> 269    68    infected         1 3.680867e+05
#> 270    69    infected         1 3.771923e+05
#> 271    70    infected         1 3.864607e+05
#> 272    71    infected         1 3.958915e+05
#> 273    72    infected         1 4.054839e+05
#> 274    73    infected         1 4.152369e+05
#> 275    74    infected         1 4.251492e+05
#> 276    75    infected         1 4.352194e+05
#> 277    76    infected         1 4.454457e+05
#> 278    77    infected         1 4.558260e+05
#> 279    78    infected         1 4.663580e+05
#> 280    79    infected         1 4.770391e+05
#> 281    80    infected         1 4.878663e+05
#> 282    81    infected         1 5.198378e+05
#> 283    82    infected         1 5.537146e+05
#> 284    83    infected         1 5.895842e+05
#> 285    84    infected         1 6.275342e+05
#> 286    85    infected         1 6.676520e+05
#> 287    86    infected         1 7.100242e+05
#> 288    87    infected         1 7.547354e+05
#> 289    88    infected         1 8.018678e+05
#> 290    89    infected         1 8.514995e+05
#> 291    90    infected         1 9.037043e+05
#> 292    91    infected         1 9.585496e+05
#> 293    92    infected         1 1.016096e+06
#> 294    93    infected         1 1.076393e+06
#> 295    94    infected         1 1.139484e+06
#> 296    95    infected         1 1.205395e+06
#> 297    96    infected         1 1.274143e+06
#> 298    97    infected         1 1.345725e+06
#> 299    98    infected         1 1.420124e+06
#> 2100   99    infected         1 1.497300e+06
#> 2101  100    infected         1 1.577195e+06
#> 2102  101    infected         1 1.658123e+06
#> 2103  102    infected         1 1.738105e+06
#> 2104  103    infected         1 1.816610e+06
#> 2105  104    infected         1 1.893099e+06
#> 2106  105    infected         1 1.967033e+06
#> 2107  106    infected         1 2.037880e+06
#> 2108  107    infected         1 2.105123e+06
#> 2109  108    infected         1 2.168269e+06
#> 2110  109    infected         1 2.226852e+06
#> 2111  110    infected         1 2.280448e+06
#> 2112  111    infected         1 2.328671e+06
#> 2113  112    infected         1 2.371187e+06
#> 2114  113    infected         1 2.407717e+06
#> 2115  114    infected         1 2.438036e+06
#> 2116  115    infected         1 2.461981e+06
#> 2117  116    infected         1 2.479448e+06
#> 2118  117    infected         1 2.490396e+06
#> 2119  118    infected         1 2.494843e+06
#> 2120  119    infected         1 2.492865e+06
#> 2121  120    infected         1 2.484594e+06
#> 2122  121    infected         1 2.470213e+06
#> 2123  122    infected         1 2.449952e+06
#> 2124  123    infected         1 2.424081e+06
#> 2125  124    infected         1 2.392910e+06
#> 2126  125    infected         1 2.356776e+06
#> 2127  126    infected         1 2.316042e+06
#> 2128  127    infected         1 2.271093e+06
#> 2129  128    infected         1 2.222322e+06
#> 2130  129    infected         1 2.170135e+06
#> 2131  130    infected         1 2.114939e+06
#> 2132  131    infected         1 2.057138e+06
#> 2133  132    infected         1 1.997132e+06
#> 2134  133    infected         1 1.935310e+06
#> 2135  134    infected         1 1.872050e+06
#> 2136  135    infected         1 1.807712e+06
#> 2137  136    infected         1 1.742639e+06
#> 2138  137    infected         1 1.677154e+06
#> 2139  138    infected         1 1.611558e+06
#> 2140  139    infected         1 1.546132e+06
#> 2141  140    infected         1 1.481132e+06
#> 2142  141    infected         1 1.416793e+06
#> 2143  142    infected         1 1.353325e+06
#> 2144  143    infected         1 1.290917e+06
#> 2145  144    infected         1 1.229735e+06
#> 2146  145    infected         1 1.169926e+06
#> 2147  146    infected         1 1.111614e+06
#> 2148  147    infected         1 1.054904e+06
#> 2149  148    infected         1 9.998848e+05
#> 2150  149    infected         1 9.466261e+05
#> 2151  150    infected         1 8.951825e+05
#> 2152  151    infected         1 8.455940e+05
#> 2153  152    infected         1 7.978870e+05
#> 2154  153    infected         1 7.520758e+05
#> 2155  154    infected         1 7.081634e+05
#> 2156  155    infected         1 6.661431e+05
#> 2157  156    infected         1 6.259992e+05
#> 2158  157    infected         1 5.877085e+05
#> 2159  158    infected         1 5.512405e+05
#> 2160  159    infected         1 5.165591e+05
#> 2161  160    infected         1 4.836232e+05
#> 2162  161    infected         1 4.523874e+05
#> 2163  162    infected         1 4.228026e+05
#> 2164  163    infected         1 3.948172e+05
#> 2165  164    infected         1 3.683772e+05
#> 2166  165    infected         1 3.434269e+05
#> 2167  166    infected         1 3.199096e+05
#> 2168  167    infected         1 2.977679e+05
#> 2169  168    infected         1 2.769439e+05
#> 2170  169    infected         1 2.573802e+05
#> 2171  170    infected         1 2.390193e+05
#> 2172  171    infected         1 2.218046e+05
#> 2173  172    infected         1 2.056805e+05
#> 2174  173    infected         1 1.905924e+05
#> 2175  174    infected         1 1.764868e+05
#> 2176  175    infected         1 1.633120e+05
#> 2177  176    infected         1 1.510174e+05
#> 2178  177    infected         1 1.395543e+05
#> 2179  178    infected         1 1.288758e+05
#> 2180  179    infected         1 1.189364e+05
#> 2181  180    infected         1 1.096926e+05
#> 2182  181    infected         1 1.011028e+05
#> 2183  182    infected         1 9.312690e+04
#> 2184  183    infected         1 8.572693e+04
#> 2185  184    infected         1 7.886654e+04
#> 2186  185    infected         1 7.251118e+04
#> 2187  186    infected         1 6.662803e+04
#> 2188  187    infected         1 6.118599e+04
#> 2189  188    infected         1 5.615559e+04
#> 2190  189    infected         1 5.150899e+04
#> 2191  190    infected         1 4.721989e+04
#> 2192  191    infected         1 4.326350e+04
#> 2193  192    infected         1 3.961647e+04
#> 2194  193    infected         1 3.625686e+04
#> 2195  194    infected         1 3.316405e+04
#> 2196  195    infected         1 3.031868e+04
#> 2197  196    infected         1 2.770265e+04
#> 2198  197    infected         1 2.529897e+04
#> 2199  198    infected         1 2.309178e+04
#> 2200  199    infected         1 2.106626e+04
#> 2201  200    infected         1 1.920859e+04
#> 31      0   recovered         1 0.000000e+00
#> 32      1   recovered         1 9.921437e+02
#> 33      2   recovered         1 2.057719e+03
#> 34      3   recovered         1 3.202155e+03
#> 35      4   recovered         1 4.431282e+03
#> 36      5   recovered         1 5.751359e+03
#> 37      6   recovered         1 7.169106e+03
#> 38      7   recovered         1 8.691740e+03
#> 39      8   recovered         1 1.032701e+04
#> 310     9   recovered         1 1.208323e+04
#> 311    10   recovered         1 1.396934e+04
#> 312    11   recovered         1 1.599491e+04
#> 313    12   recovered         1 1.817025e+04
#> 314    13   recovered         1 2.050640e+04
#> 315    14   recovered         1 2.301522e+04
#> 316    15   recovered         1 2.570944e+04
#> 317    16   recovered         1 2.860273e+04
#> 318    17   recovered         1 3.170974e+04
#> 319    18   recovered         1 3.504623e+04
#> 320    19   recovered         1 3.862907e+04
#> 321    20   recovered         1 4.247640e+04
#> 322    21   recovered         1 4.660765e+04
#> 323    22   recovered         1 5.104370e+04
#> 324    23   recovered         1 5.580694e+04
#> 325    24   recovered         1 6.092137e+04
#> 326    25   recovered         1 6.641276e+04
#> 327    26   recovered         1 7.230874e+04
#> 328    27   recovered         1 7.863892e+04
#> 329    28   recovered         1 8.543510e+04
#> 330    29   recovered         1 9.273132e+04
#> 331    30   recovered         1 1.005641e+05
#> 332    31   recovered         1 1.089726e+05
#> 333    32   recovered         1 1.179988e+05
#> 334    33   recovered         1 1.276876e+05
#> 335    34   recovered         1 1.380872e+05
#> 336    35   recovered         1 1.492491e+05
#> 337    36   recovered         1 1.612287e+05
#> 338    37   recovered         1 1.740850e+05
#> 339    38   recovered         1 1.878814e+05
#> 340    39   recovered         1 2.026857e+05
#> 341    40   recovered         1 2.185704e+05
#> 342    41   recovered         1 2.356129e+05
#> 343    42   recovered         1 2.538963e+05
#> 344    43   recovered         1 2.735091e+05
#> 345    44   recovered         1 2.945461e+05
#> 346    45   recovered         1 3.171083e+05
#> 347    46   recovered         1 3.413039e+05
#> 348    47   recovered         1 3.672480e+05
#> 349    48   recovered         1 3.950636e+05
#> 350    49   recovered         1 4.248816e+05
#> 351    50   recovered         1 4.568418e+05
#> 352    51   recovered         1 4.903688e+05
#> 353    52   recovered         1 5.248032e+05
#> 354    53   recovered         1 5.601661e+05
#> 355    54   recovered         1 5.964784e+05
#> 356    55   recovered         1 6.337616e+05
#> 357    56   recovered         1 6.720373e+05
#> 358    57   recovered         1 7.113274e+05
#> 359    58   recovered         1 7.516540e+05
#> 360    59   recovered         1 7.930393e+05
#> 361    60   recovered         1 8.355059e+05
#> 362    61   recovered         1 8.790764e+05
#> 363    62   recovered         1 9.237736e+05
#> 364    63   recovered         1 9.696206e+05
#> 365    64   recovered         1 1.016640e+06
#> 366    65   recovered         1 1.064856e+06
#> 367    66   recovered         1 1.114291e+06
#> 368    67   recovered         1 1.164968e+06
#> 369    68   recovered         1 1.216911e+06
#> 370    69   recovered         1 1.270143e+06
#> 371    70   recovered         1 1.324688e+06
#> 372    71   recovered         1 1.380568e+06
#> 373    72   recovered         1 1.437808e+06
#> 374    73   recovered         1 1.496429e+06
#> 375    74   recovered         1 1.556454e+06
#> 376    75   recovered         1 1.617907e+06
#> 377    76   recovered         1 1.680810e+06
#> 378    77   recovered         1 1.745185e+06
#> 379    78   recovered         1 1.811053e+06
#> 380    79   recovered         1 1.878437e+06
#> 381    80   recovered         1 1.947357e+06
#> 382    81   recovered         1 2.019314e+06
#> 383    82   recovered         1 2.095973e+06
#> 384    83   recovered         1 2.177613e+06
#> 385    84   recovered         1 2.264525e+06
#> 386    85   recovered         1 2.357012e+06
#> 387    86   recovered         1 2.455390e+06
#> 388    87   recovered         1 2.559987e+06
#> 389    88   recovered         1 2.671144e+06
#> 390    89   recovered         1 2.789211e+06
#> 391    90   recovered         1 2.914552e+06
#> 392    91   recovered         1 3.047538e+06
#> 393    92   recovered         1 3.188552e+06
#> 394    93   recovered         1 3.337982e+06
#> 395    94   recovered         1 3.496226e+06
#> 396    95   recovered         1 3.663683e+06
#> 397    96   recovered         1 3.840760e+06
#> 398    97   recovered         1 4.027859e+06
#> 399    98   recovered         1 4.225387e+06
#> 3100   99   recovered         1 4.433741e+06
#> 3101  100   recovered         1 4.653316e+06
#> 3102  101   recovered         1 4.884418e+06
#> 3103  102   recovered         1 5.127021e+06
#> 3104  103   recovered         1 5.380950e+06
#> 3105  104   recovered         1 5.645956e+06
#> 3106  105   recovered         1 5.921713e+06
#> 3107  106   recovered         1 6.207818e+06
#> 3108  107   recovered         1 6.503793e+06
#> 3109  108   recovered         1 6.809087e+06
#> 3110  109   recovered         1 7.123081e+06
#> 3111  110   recovered         1 7.445093e+06
#> 3112  111   recovered         1 7.774382e+06
#> 3113  112   recovered         1 8.110156e+06
#> 3114  113   recovered         1 8.451579e+06
#> 3115  114   recovered         1 8.797779e+06
#> 3116  115   recovered         1 9.147857e+06
#> 3117  116   recovered         1 9.500893e+06
#> 3118  117   recovered         1 9.855960e+06
#> 3119  118   recovered         1 1.021213e+07
#> 3120  119   recovered         1 1.056847e+07
#> 3121  120   recovered         1 1.092407e+07
#> 3122  121   recovered         1 1.127806e+07
#> 3123  122   recovered         1 1.162957e+07
#> 3124  123   recovered         1 1.197778e+07
#> 3125  124   recovered         1 1.232191e+07
#> 3126  125   recovered         1 1.266123e+07
#> 3127  126   recovered         1 1.299506e+07
#> 3128  127   recovered         1 1.332276e+07
#> 3129  128   recovered         1 1.364376e+07
#> 3130  129   recovered         1 1.395754e+07
#> 3131  130   recovered         1 1.426365e+07
#> 3132  131   recovered         1 1.456169e+07
#> 3133  132   recovered         1 1.485130e+07
#> 3134  133   recovered         1 1.513221e+07
#> 3135  134   recovered         1 1.540418e+07
#> 3136  135   recovered         1 1.566703e+07
#> 3137  136   recovered         1 1.592063e+07
#> 3138  137   recovered         1 1.616491e+07
#> 3139  138   recovered         1 1.639981e+07
#> 3140  139   recovered         1 1.662536e+07
#> 3141  140   recovered         1 1.684159e+07
#> 3142  141   recovered         1 1.704857e+07
#> 3143  142   recovered         1 1.724643e+07
#> 3144  143   recovered         1 1.743529e+07
#> 3145  144   recovered         1 1.761532e+07
#> 3146  145   recovered         1 1.778671e+07
#> 3147  146   recovered         1 1.794965e+07
#> 3148  147   recovered         1 1.810439e+07
#> 3149  148   recovered         1 1.825114e+07
#> 3150  149   recovered         1 1.839015e+07
#> 3151  150   recovered         1 1.852169e+07
#> 3152  151   recovered         1 1.864601e+07
#> 3153  152   recovered         1 1.876337e+07
#> 3154  153   recovered         1 1.887406e+07
#> 3155  154   recovered         1 1.897834e+07
#> 3156  155   recovered         1 1.907649e+07
#> 3157  156   recovered         1 1.916876e+07
#> 3158  157   recovered         1 1.925543e+07
#> 3159  158   recovered         1 1.933676e+07
#> 3160  159   recovered         1 1.941301e+07
#> 3161  160   recovered         1 1.948443e+07
#> 3162  161   recovered         1 1.955127e+07
#> 3163  162   recovered         1 1.961377e+07
#> 3164  163   recovered         1 1.967215e+07
#> 3165  164   recovered         1 1.972665e+07
#> 3166  165   recovered         1 1.977747e+07
#> 3167  166   recovered         1 1.982484e+07
#> 3168  167   recovered         1 1.986894e+07
#> 3169  168   recovered         1 1.990997e+07
#> 3170  169   recovered         1 1.994813e+07
#> 3171  170   recovered         1 1.998357e+07
#> 3172  171   recovered         1 2.001647e+07
#> 3173  172   recovered         1 2.004699e+07
#> 3174  173   recovered         1 2.007529e+07
#> 3175  174   recovered         1 2.010150e+07
#> 3176  175   recovered         1 2.012576e+07
#> 3177  176   recovered         1 2.014820e+07
#> 3178  177   recovered         1 2.016894e+07
#> 3179  178   recovered         1 2.018811e+07
#> 3180  179   recovered         1 2.020580e+07
#> 3181  180   recovered         1 2.022212e+07
#> 3182  181   recovered         1 2.023717e+07
#> 3183  182   recovered         1 2.025104e+07
#> 3184  183   recovered         1 2.026381e+07
#> 3185  184   recovered         1 2.027556e+07
#> 3186  185   recovered         1 2.028636e+07
#> 3187  186   recovered         1 2.029630e+07
#> 3188  187   recovered         1 2.030542e+07
#> 3189  188   recovered         1 2.031380e+07
#> 3190  189   recovered         1 2.032149e+07
#> 3191  190   recovered         1 2.032853e+07
#> 3192  191   recovered         1 2.033499e+07
#> 3193  192   recovered         1 2.034091e+07
#> 3194  193   recovered         1 2.034632e+07
#> 3195  194   recovered         1 2.035128e+07
#> 3196  195   recovered         1 2.035581e+07
#> 3197  196   recovered         1 2.035995e+07
#> 3198  197   recovered         1 2.036374e+07
#> 3199  198   recovered         1 2.036719e+07
#> 3200  199   recovered         1 2.037034e+07
#> 3201  200   recovered         1 2.037322e+07
#> 41      0  vaccinated         1 0.000000e+00
#> 42      1  vaccinated         1 0.000000e+00
#> 43      2  vaccinated         1 0.000000e+00
#> 44      3  vaccinated         1 0.000000e+00
#> 45      4  vaccinated         1 0.000000e+00
#> 46      5  vaccinated         1 0.000000e+00
#> 47      6  vaccinated         1 0.000000e+00
#> 48      7  vaccinated         1 0.000000e+00
#> 49      8  vaccinated         1 0.000000e+00
#> 410     9  vaccinated         1 0.000000e+00
#> 411    10  vaccinated         1 0.000000e+00
#> 412    11  vaccinated         1 0.000000e+00
#> 413    12  vaccinated         1 0.000000e+00
#> 414    13  vaccinated         1 0.000000e+00
#> 415    14  vaccinated         1 0.000000e+00
#> 416    15  vaccinated         1 0.000000e+00
#> 417    16  vaccinated         1 0.000000e+00
#> 418    17  vaccinated         1 0.000000e+00
#> 419    18  vaccinated         1 0.000000e+00
#> 420    19  vaccinated         1 0.000000e+00
#> 421    20  vaccinated         1 0.000000e+00
#> 422    21  vaccinated         1 0.000000e+00
#> 423    22  vaccinated         1 0.000000e+00
#> 424    23  vaccinated         1 0.000000e+00
#> 425    24  vaccinated         1 0.000000e+00
#> 426    25  vaccinated         1 0.000000e+00
#> 427    26  vaccinated         1 0.000000e+00
#> 428    27  vaccinated         1 0.000000e+00
#> 429    28  vaccinated         1 0.000000e+00
#> 430    29  vaccinated         1 0.000000e+00
#> 431    30  vaccinated         1 0.000000e+00
#> 432    31  vaccinated         1 0.000000e+00
#> 433    32  vaccinated         1 0.000000e+00
#> 434    33  vaccinated         1 0.000000e+00
#> 435    34  vaccinated         1 0.000000e+00
#> 436    35  vaccinated         1 0.000000e+00
#> 437    36  vaccinated         1 0.000000e+00
#> 438    37  vaccinated         1 0.000000e+00
#> 439    38  vaccinated         1 0.000000e+00
#> 440    39  vaccinated         1 0.000000e+00
#> 441    40  vaccinated         1 0.000000e+00
#> 442    41  vaccinated         1 0.000000e+00
#> 443    42  vaccinated         1 0.000000e+00
#> 444    43  vaccinated         1 0.000000e+00
#> 445    44  vaccinated         1 0.000000e+00
#> 446    45  vaccinated         1 0.000000e+00
#> 447    46  vaccinated         1 0.000000e+00
#> 448    47  vaccinated         1 0.000000e+00
#> 449    48  vaccinated         1 0.000000e+00
#> 450    49  vaccinated         1 0.000000e+00
#> 451    50  vaccinated         1 0.000000e+00
#> 452    51  vaccinated         1 0.000000e+00
#> 453    52  vaccinated         1 0.000000e+00
#> 454    53  vaccinated         1 0.000000e+00
#> 455    54  vaccinated         1 0.000000e+00
#> 456    55  vaccinated         1 0.000000e+00
#> 457    56  vaccinated         1 0.000000e+00
#> 458    57  vaccinated         1 0.000000e+00
#> 459    58  vaccinated         1 0.000000e+00
#> 460    59  vaccinated         1 0.000000e+00
#> 461    60  vaccinated         1 0.000000e+00
#> 462    61  vaccinated         1 0.000000e+00
#> 463    62  vaccinated         1 0.000000e+00
#> 464    63  vaccinated         1 0.000000e+00
#> 465    64  vaccinated         1 0.000000e+00
#> 466    65  vaccinated         1 0.000000e+00
#> 467    66  vaccinated         1 0.000000e+00
#> 468    67  vaccinated         1 0.000000e+00
#> 469    68  vaccinated         1 0.000000e+00
#> 470    69  vaccinated         1 0.000000e+00
#> 471    70  vaccinated         1 0.000000e+00
#> 472    71  vaccinated         1 0.000000e+00
#> 473    72  vaccinated         1 0.000000e+00
#> 474    73  vaccinated         1 0.000000e+00
#> 475    74  vaccinated         1 0.000000e+00
#> 476    75  vaccinated         1 0.000000e+00
#> 477    76  vaccinated         1 0.000000e+00
#> 478    77  vaccinated         1 0.000000e+00
#> 479    78  vaccinated         1 0.000000e+00
#> 480    79  vaccinated         1 0.000000e+00
#> 481    80  vaccinated         1 0.000000e+00
#> 482    81  vaccinated         1 0.000000e+00
#> 483    82  vaccinated         1 0.000000e+00
#> 484    83  vaccinated         1 0.000000e+00
#> 485    84  vaccinated         1 0.000000e+00
#> 486    85  vaccinated         1 0.000000e+00
#> 487    86  vaccinated         1 0.000000e+00
#> 488    87  vaccinated         1 0.000000e+00
#> 489    88  vaccinated         1 0.000000e+00
#> 490    89  vaccinated         1 0.000000e+00
#> 491    90  vaccinated         1 0.000000e+00
#> 492    91  vaccinated         1 0.000000e+00
#> 493    92  vaccinated         1 0.000000e+00
#> 494    93  vaccinated         1 0.000000e+00
#> 495    94  vaccinated         1 0.000000e+00
#> 496    95  vaccinated         1 0.000000e+00
#> 497    96  vaccinated         1 0.000000e+00
#> 498    97  vaccinated         1 0.000000e+00
#> 499    98  vaccinated         1 0.000000e+00
#> 4100   99  vaccinated         1 0.000000e+00
#> 4101  100  vaccinated         1 9.835326e-07
#> 4102  101  vaccinated         1 6.031206e+05
#> 4103  102  vaccinated         1 1.197082e+06
#> 4104  103  vaccinated         1 1.781874e+06
#> 4105  104  vaccinated         1 2.357494e+06
#> 4106  105  vaccinated         1 2.923946e+06
#> 4107  106  vaccinated         1 3.481246e+06
#> 4108  107  vaccinated         1 4.029417e+06
#> 4109  108  vaccinated         1 4.568492e+06
#> 4110  109  vaccinated         1 5.098516e+06
#> 4111  110  vaccinated         1 5.619543e+06
#> 4112  111  vaccinated         1 6.131638e+06
#> 4113  112  vaccinated         1 6.634876e+06
#> 4114  113  vaccinated         1 7.129344e+06
#> 4115  114  vaccinated         1 7.615137e+06
#> 4116  115  vaccinated         1 8.092362e+06
#> 4117  116  vaccinated         1 8.561133e+06
#> 4118  117  vaccinated         1 9.021575e+06
#> 4119  118  vaccinated         1 9.473820e+06
#> 4120  119  vaccinated         1 9.918007e+06
#> 4121  120  vaccinated         1 1.035428e+07
#> 4122  121  vaccinated         1 1.078280e+07
#> 4123  122  vaccinated         1 1.120371e+07
#> 4124  123  vaccinated         1 1.161719e+07
#> 4125  124  vaccinated         1 1.202339e+07
#> 4126  125  vaccinated         1 1.242248e+07
#> 4127  126  vaccinated         1 1.281464e+07
#> 4128  127  vaccinated         1 1.320002e+07
#> 4129  128  vaccinated         1 1.357882e+07
#> 4130  129  vaccinated         1 1.395118e+07
#> 4131  130  vaccinated         1 1.431730e+07
#> 4132  131  vaccinated         1 1.467732e+07
#> 4133  132  vaccinated         1 1.503143e+07
#> 4134  133  vaccinated         1 1.537978e+07
#> 4135  134  vaccinated         1 1.572253e+07
#> 4136  135  vaccinated         1 1.605986e+07
#> 4137  136  vaccinated         1 1.639189e+07
#> 4138  137  vaccinated         1 1.671880e+07
#> 4139  138  vaccinated         1 1.704072e+07
#> 4140  139  vaccinated         1 1.735781e+07
#> 4141  140  vaccinated         1 1.767018e+07
#> 4142  141  vaccinated         1 1.797799e+07
#> 4143  142  vaccinated         1 1.828136e+07
#> 4144  143  vaccinated         1 1.858041e+07
#> 4145  144  vaccinated         1 1.887527e+07
#> 4146  145  vaccinated         1 1.916604e+07
#> 4147  146  vaccinated         1 1.945285e+07
#> 4148  147  vaccinated         1 1.973579e+07
#> 4149  148  vaccinated         1 2.001498e+07
#> 4150  149  vaccinated         1 2.029051e+07
#> 4151  150  vaccinated         1 2.056247e+07
#> 4152  151  vaccinated         1 2.083095e+07
#> 4153  152  vaccinated         1 2.109604e+07
#> 4154  153  vaccinated         1 2.135783e+07
#> 4155  154  vaccinated         1 2.161639e+07
#> 4156  155  vaccinated         1 2.187180e+07
#> 4157  156  vaccinated         1 2.212412e+07
#> 4158  157  vaccinated         1 2.237344e+07
#> 4159  158  vaccinated         1 2.261981e+07
#> 4160  159  vaccinated         1 2.286331e+07
#> 4161  160  vaccinated         1 2.310398e+07
#> 4162  161  vaccinated         1 2.334188e+07
#> 4163  162  vaccinated         1 2.357708e+07
#> 4164  163  vaccinated         1 2.380963e+07
#> 4165  164  vaccinated         1 2.403957e+07
#> 4166  165  vaccinated         1 2.426695e+07
#> 4167  166  vaccinated         1 2.449183e+07
#> 4168  167  vaccinated         1 2.471424e+07
#> 4169  168  vaccinated         1 2.493422e+07
#> 4170  169  vaccinated         1 2.515183e+07
#> 4171  170  vaccinated         1 2.536709e+07
#> 4172  171  vaccinated         1 2.558004e+07
#> 4173  172  vaccinated         1 2.579073e+07
#> 4174  173  vaccinated         1 2.599919e+07
#> 4175  174  vaccinated         1 2.620544e+07
#> 4176  175  vaccinated         1 2.640953e+07
#> 4177  176  vaccinated         1 2.661148e+07
#> 4178  177  vaccinated         1 2.681133e+07
#> 4179  178  vaccinated         1 2.700909e+07
#> 4180  179  vaccinated         1 2.720481e+07
#> 4181  180  vaccinated         1 2.739851e+07
#> 4182  181  vaccinated         1 2.759022e+07
#> 4183  182  vaccinated         1 2.777995e+07
#> 4184  183  vaccinated         1 2.796774e+07
#> 4185  184  vaccinated         1 2.815361e+07
#> 4186  185  vaccinated         1 2.833759e+07
#> 4187  186  vaccinated         1 2.851969e+07
#> 4188  187  vaccinated         1 2.869994e+07
#> 4189  188  vaccinated         1 2.887837e+07
#> 4190  189  vaccinated         1 2.905498e+07
#> 4191  190  vaccinated         1 2.922982e+07
#> 4192  191  vaccinated         1 2.940288e+07
#> 4193  192  vaccinated         1 2.957420e+07
#> 4194  193  vaccinated         1 2.974379e+07
#> 4195  194  vaccinated         1 2.991168e+07
#> 4196  195  vaccinated         1 3.007788e+07
#> 4197  196  vaccinated         1 3.024240e+07
#> 4198  197  vaccinated         1 3.040528e+07
#> 4199  198  vaccinated         1 3.056652e+07
#> 4200  199  vaccinated         1 3.072615e+07
#> 4201  200  vaccinated         1 3.088418e+07