Model an epidemic

Simulate an epidemic using a deterministic, compartmental epidemic model from the package's library of epidemic models. The only option currently available is a SEIR-V model, with the compartments "susceptible", "exposed", "infectious", "recovered", and "vaccinated".

Usage

epidemic(model = "default", ...)

Arguments

model

A string for the epidemic model. The only currently supported option is "default", for the default SEIR-V model.

...

Arguments to the model specified by model. See Details for more on the supported arguments.

Value

A data.frame with the columns "time", "compartment", "age_group", "value". The compartments correspond to the compartments of the model chosen with model. The current default model has the compartments "susceptible", "exposed", "infectious", "recovered", and "vaccinated".

Details

Arguments passed in ... may differ depending on the model specified in model. The default model (model = "default") takes the following arguments.

population An object of the population class, which holds a population contact matrix, a demography vector, and the initial conditions of each demographic group. See population().

r0 The reproductive number of the infection, \(R_0\). Must be a number or a vector of numbers of the same length as the number of demographic groups, depending on the model being implemented. The 'default' model requires a single number.

preinfectious_period The mean infectious period. Must be a number or a vector of numbers of the same length as the number of demographic groups, depending on the model being implemented. The 'default' model requires a single number.

infectious_period The mean infections period. Must be a number or a vector of numbers of the same length as the number of demographic groups, depending on the model being implemented. The 'default' model requires a single number.

intervention A non-pharmaceutical intervention applied to the population during the epidemic. See intervention(). This is an optional argument in the default model.

vaccination A vaccination regime followed during the course of the epidemic, with a start and end time, and age-specific effect on the transition of individuals from susceptible to vaccinated. See vaccination(). This is an optional argument in the default model.

time_end The maximum number of timesteps over which to run the model.

increment The size of the time increment.

Examples

# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# run epidemic simulation with no vaccination or intervention
epidemic(
  model = "default",
  population = uk_population,
  r0 = 1.5,
  preinfectious_period = 3,
  infectious_period = 7,
  time_end = 200,
  increment = 1
)
#> Intervention: `time_end` is not greater than `time_begin`
#> Vaccination: some `time_end`s are not greater than `time_begin`s
#>     susceptible_1   exposed_1 infectious_1  recovered_1 vaccinated_1 time
#> 1        66993300    6700.000        0.000 0.000000e+00            0    0
#> 2        66993094    4986.052     1782.694 1.375334e+02            0    1
#> 3        66992580    4015.446     2924.359 4.797614e+02            0    2
#> 4        66991867    3488.386     3689.298 9.556865e+02            0    3
#> 5        66991015    3226.973     4234.621 1.523722e+03            0    4
#> 6        66990061    3125.685     4653.797 2.159769e+03            0    5
#> 7        66989025    3121.984     5002.523 2.850141e+03            0    6
#> 8        66987920    3178.853     5314.073 3.587361e+03            0    7
#> 9        66986749    3274.438     5608.421 4.367670e+03            0    8
#> 10       66985517    3395.890     5897.654 5.189548e+03            0    9
#> 11       66984222    3535.713     6189.183 6.052837e+03            0   10
#> 12       66982865    3689.598     6487.658 6.958225e+03            0   11
#> 13       66981442    3855.132     6796.099 7.906934e+03            0   12
#> 14       66979952    4031.035     7116.571 8.900543e+03            0   13
#> 15       66978392    4216.709     7450.584 9.940886e+03            0   14
#> 16       66976759    4411.964     7799.331 1.102998e+04            0   15
#> 17       66975049    4616.864     8163.832 1.217002e+04            0   16
#> 18       66973260    4831.631     8545.018 1.336330e+04            0   17
#> 19       66971387    5056.587     8943.780 1.461229e+04            0   18
#> 20       66969427    5292.125     9361.006 1.591955e+04            0   19
#> 21       66967376    5538.688     9797.593 1.728778e+04            0   20
#> 22       66965229    5796.757    10254.465 1.871983e+04            0   21
#> 23       66962982    6066.850    10732.576 2.021864e+04            0   22
#> 24       66960630    6349.512    11232.919 2.178733e+04            0   23
#> 25       66958169    6645.318    11756.530 2.342915e+04            0   24
#> 26       66955593    6954.873    12304.490 2.514750e+04            0   25
#> 27       66952897    7278.810    12877.928 2.694594e+04            0   26
#> 28       66950076    7617.791    13478.024 2.882818e+04            0   27
#> 29       66947123    7972.511    14106.012 3.079813e+04            0   28
#> 30       66944033    8343.694    14763.182 3.285986e+04            0   29
#> 31       66940799    8732.099    15450.883 3.501763e+04            0   30
#> 32       66937415    9138.520    16170.524 3.727592e+04            0   31
#> 33       66933873    9563.786    16923.582 3.963937e+04            0   32
#> 34       66930167   10008.764    17711.598 4.211289e+04            0   33
#> 35       66926288   10474.359    18536.184 4.470157e+04            0   34
#> 36       66922229   10961.519    19399.026 4.741076e+04            0   35
#> 37       66917981   11471.234    20301.885 5.024606e+04            0   36
#> 38       66913536   12004.539    21246.605 5.321330e+04            0   37
#> 39       66908884   12562.513    22235.109 5.631860e+04            0   38
#> 40       66904016   13146.288    23269.411 5.956836e+04            0   39
#> 41       66898922   13757.041    24351.614 6.296928e+04            0   40
#> 42       66893592   14396.008    25483.917 6.652835e+04            0   41
#> 43       66888014   15064.474    26668.616 7.025290e+04            0   42
#> 44       66882178   15763.786    27908.113 7.415057e+04            0   43
#> 45       66876070   16495.348    29204.915 7.822937e+04            0   44
#> 46       66869680   17260.627    30561.643 8.249768e+04            0   45
#> 47       66862994   18061.156    31981.035 8.696425e+04            0   46
#> 48       66855997   18898.532    33465.949 9.163824e+04            0   47
#> 49       66848677   19774.426    35019.374 9.652922e+04            0   48
#> 50       66841018   20690.580    36644.427 1.016472e+05            0   49
#> 51       66833004   21648.814    38344.367 1.070026e+05            0   50
#> 52       66824620   22651.024    40122.595 1.126064e+05            0   51
#> 53       66815848   23699.191    41982.661 1.184701e+05            0   52
#> 54       66806671   24795.381    43928.272 1.246056e+05            0   53
#> 55       66797070   25941.748    45963.298 1.310253e+05            0   54
#> 56       66787025   27140.539    48091.775 1.377424e+05            0   55
#> 57       66776518   28394.097    50317.918 1.447705e+05            0   56
#> 58       66765525   29704.863    52646.122 1.521238e+05            0   57
#> 59       66754026   31075.383    55080.972 1.598173e+05            0   58
#> 60       66741998   32508.309    57627.252 1.678665e+05            0   59
#> 61       66729416   34006.405    60289.947 1.762878e+05            0   60
#> 62       66716255   35572.549    63074.259 1.850980e+05            0   61
#> 63       66702490   37209.737    65985.608 1.943150e+05            0   62
#> 64       66688092   38921.090    69029.642 2.039574e+05            0   63
#> 65       66673034   40709.855    72212.249 2.140444e+05            0   64
#> 66       66657285   42579.411    75539.561 2.245963e+05            0   65
#> 67       66640814   44533.273    79017.965 2.356343e+05            0   66
#> 68       66623590   46575.096    82654.111 2.471804e+05            0   67
#> 69       66605579   48708.679    86454.924 2.592576e+05            0   68
#> 70       66586744   50937.971    90427.609 2.718900e+05            0   69
#> 71       66567051   53267.073    94579.662 2.851026e+05            0   70
#> 72       66546459   55700.245    98918.882 2.989216e+05            0   71
#> 73       66524930   58241.907   103453.378 3.133744e+05            0   72
#> 74       66502422   60896.647   108191.578 3.284895e+05            0   73
#> 75       66478892   63669.221   113142.241 3.442964e+05            0   74
#> 76       66454295   66564.563   118314.464 3.608264e+05            0   75
#> 77       66428583   69587.780   123717.694 3.781116e+05            0   76
#> 78       66401708   72744.166   129361.737 3.961857e+05            0   77
#> 79       66373620   76039.196   135256.765 4.150840e+05            0   78
#> 80       66344265   79478.536   141413.328 4.348430e+05            0   79
#> 81       66313589   83068.043   147842.362 4.555008e+05            0   80
#> 82       66281534   86813.767   154555.196 4.770972e+05            0   81
#> 83       66248041   90721.954   161563.562 4.996735e+05            0   82
#> 84       66213049   94799.049   168879.602 5.232728e+05            0   83
#> 85       66176492   99051.694   176515.875 5.479400e+05            0   84
#> 86       66138306  103486.731   184485.362 5.737218e+05            0   85
#> 87       66098421  108111.202   192801.475 6.006666e+05            0   86
#> 88       66056765  112932.345   201478.056 6.288251e+05            0   87
#> 89       66013263  117957.595   210529.387 6.582496e+05            0   88
#> 90       65967840  123194.583   219970.188 6.889949e+05            0   89
#> 91       65920416  128651.127   229815.621 7.211175e+05            0   90
#> 92       65870907  134335.232   240081.285 7.546765e+05            0   91
#> 93       65819229  140255.079   250783.220 7.897330e+05            0   92
#> 94       65765293  146419.023   261937.898 8.263504e+05            0   93
#> 95       65709007  152835.581   273562.220 8.645947e+05            0   94
#> 96       65650279  159513.420   285673.508 9.045342e+05            0   95
#> 97       65589009  166461.348   298289.491 9.462397e+05            0   96
#> 98       65525099  173688.297   311428.296 9.897847e+05            0   97
#> 99       65458443  181203.307   325108.425 1.035245e+06            0   98
#> 100      65388936  189015.508   339348.746 1.082699e+06            0   99
#> 101      65316468  197134.099   354168.456 1.132229e+06            0  100
#> 102      65240926  205568.323   369587.067 1.183919e+06            0  101
#> 103      65162193  214327.442   385624.366 1.237855e+06            0  102
#> 104      65080151  223420.709   402300.381 1.294128e+06            0  103
#> 105      64994678  232857.331   419635.344 1.352829e+06            0  104
#> 106      64905648  242646.438   437649.640 1.414056e+06            0  105
#> 107      64812934  252797.042   456363.761 1.477906e+06            0  106
#> 108      64716404  263317.996   475798.244 1.544480e+06            0  107
#> 109      64615925  274217.944   495973.612 1.613883e+06            0  108
#> 110      64511362  285505.275   516910.299 1.686223e+06            0  109
#> 111      64402574  297188.065   538628.579 1.761609e+06            0  110
#> 112      64289423  309274.022   561148.477 1.840155e+06            0  111
#> 113      64171764  321770.422   584489.681 1.921976e+06            0  112
#> 114      64049453  334684.039   608671.438 2.007192e+06            0  113
#> 115      63922343  348021.080   633712.450 2.095923e+06            0  114
#> 116      63790288  361787.104   659630.757 2.188294e+06            0  115
#> 117      63653138  375986.946   686443.610 2.284432e+06            0  116
#> 118      63510744  390624.628   714167.339 2.384464e+06            0  117
#> 119      63362956  405703.278   742817.209 2.488524e+06            0  118
#> 120      63209625  421225.031   772407.269 2.596743e+06            0  119
#> 121      63050602  437190.937   802950.192 2.709257e+06            0  120
#> 122      62885739  453600.864   834457.104 2.826203e+06            0  121
#> 123      62714890  470453.393   866937.406 2.947719e+06            0  122
#> 124      62537910  487745.713   900398.591 3.073946e+06            0  123
#> 125      62354657  505473.522   934846.043 3.205023e+06            0  124
#> 126      62164994  523630.914   970282.845 3.341092e+06            0  125
#> 127      61968786  542210.272  1006709.562 3.482294e+06            0  126
#> 128      61765904  561202.169  1044124.034 3.628770e+06            0  127
#> 129      61556222  580595.253  1082521.159 3.780662e+06            0  128
#> 130      61339623  600376.155  1121892.668 3.938108e+06            0  129
#> 131      61115995  620529.385  1162226.904 4.101248e+06            0  130
#> 132      60885236  641037.240  1203508.604 4.270218e+06            0  131
#> 133      60647250  661879.720  1245718.672 4.445152e+06            0  132
#> 134      60401951  683034.452  1288833.969 4.626181e+06            0  133
#> 135      60149264  704476.615  1332827.096 4.813432e+06            0  134
#> 136      59889126  726178.891  1377666.198 5.007029e+06            0  135
#> 137      59621484  748111.418  1423314.771 5.207089e+06            0  136
#> 138      59346300  770241.761  1469731.487 5.413727e+06            0  137
#> 139      59063548  792534.900  1516870.032 5.627047e+06            0  138
#> 140      58773218  814953.239  1564678.971 5.847150e+06            0  139
#> 141      58475314  837456.624  1613101.631 6.074127e+06            0  140
#> 142      58169859  860002.401  1662076.009 6.308062e+06            0  141
#> 143      57856891  882545.473  1711534.718 6.549029e+06            0  142
#> 144      57536465  905038.407  1761404.962 6.797092e+06            0  143
#> 145      57208656  927431.540  1811608.539 7.052303e+06            0  144
#> 146      56873559  949673.134  1862061.901 7.314706e+06            0  145
#> 147      56531285  971709.542  1912676.242 7.584329e+06            0  146
#> 148      56181969  993485.404  1963357.636 7.861188e+06            0  147
#> 149      55825762 1014943.875  2014007.223 8.145286e+06            0  148
#> 150      55462840 1036026.874  2064521.442 8.436612e+06            0  149
#> 151      55093394 1056675.356  2114792.317 8.735138e+06            0  150
#> 152      54717641 1076829.616  2164707.794 9.040821e+06            0  151
#> 153      54335815 1096429.604  2214152.124 9.353603e+06            0  152
#> 154      53948170 1115415.268  2263006.307 9.673408e+06            0  153
#> 155      53554982 1133726.914  2311148.579 1.000014e+07            0  154
#> 156      53156543 1151305.570  2358454.945 1.033370e+07            0  155
#> 157      52753166 1168093.375  2404799.770 1.067394e+07            0  156
#> 158      52345180 1184033.963  2450056.396 1.102073e+07            0  157
#> 159      51932930 1199072.854  2494097.807 1.137390e+07            0  158
#> 160      51516780 1213157.845  2536797.326 1.173327e+07            0  159
#> 161      51097103 1226239.390  2578029.339 1.209863e+07            0  160
#> 162      50674290 1238270.970  2617670.032 1.246977e+07            0  161
#> 163      50248740 1249209.454  2655598.156 1.284645e+07            0  162
#> 164      49820865 1259015.426  2691695.788 1.322842e+07            0  163
#> 165      49391082 1267653.501  2725849.092 1.361542e+07            0  164
#> 166      48959818 1275092.605  2757949.075 1.400714e+07            0  165
#> 167      48527503 1281306.217  2787892.318 1.440330e+07            0  166
#> 168      48094572 1286272.588  2815581.686 1.480357e+07            0  167
#> 169      47661460 1289974.908  2840927.000 1.520764e+07            0  168
#> 170      47228602 1292401.439  2863845.669 1.561515e+07            0  169
#> 171      46796431 1293545.597  2884263.267 1.602576e+07            0  170
#> 172      46365376 1293405.998  2902114.059 1.643910e+07            0  171
#> 173      45935861 1291986.448  2917341.456 1.685481e+07            0  172
#> 174      45508303 1289295.903  2929898.404 1.727250e+07            0  173
#> 175      45083108 1285348.365  2939747.703 1.769180e+07            0  174
#> 176      44660674 1280162.753  2946862.239 1.811230e+07            0  175
#> 177      44241387 1273762.724  2951225.141 1.853363e+07            0  176
#> 178      43825617 1266176.457  2952829.860 1.895538e+07            0  177
#> 179      43413724 1257436.406  2951680.157 1.937716e+07            0  178
#> 180      43006048 1247579.017  2947790.015 1.979858e+07            0  179
#> 181      42602916 1236644.422  2941183.473 2.021926e+07            0  180
#> 182      42204636 1224676.107  2931894.376 2.063879e+07            0  181
#> 183      41811499 1211720.568  2919966.059 2.105681e+07            0  182
#> 184      41423776 1197826.942  2905450.962 2.147295e+07            0  183
#> 185      41041720 1183046.644  2888410.176 2.188682e+07            0  184
#> 186      40665565 1167432.982  2868912.943 2.229809e+07            0  185
#> 187      40295524 1151040.788  2847036.096 2.270640e+07            0  186
#> 188      39931792 1133926.042  2822863.466 2.311142e+07            0  187
#> 189      39574543 1116145.508  2796485.245 2.351283e+07            0  188
#> 190      39223932 1097756.382  2767997.329 2.391031e+07            0  189
#> 191      38880096 1078815.949  2737500.636 2.430359e+07            0  190
#> 192      38543152 1059381.267  2705100.417 2.469237e+07            0  191
#> 193      38213198 1039508.855  2670905.554 2.507639e+07            0  192
#> 194      37890315 1019254.423  2635027.873 2.545540e+07            0  193
#> 195      37574568  998672.604  2597581.448 2.582918e+07            0  194
#> 196      37266002  977816.723  2558681.939 2.619750e+07            0  195
#> 197      36964649  956738.587  2518445.935 2.656017e+07            0  196
#> 198      36670524  935488.301  2476990.330 2.691700e+07            0  197
#> 199      36383628  914114.103  2434431.728 2.726783e+07            0  198
#> 200      36103949  892662.232  2390885.880 2.761250e+07            0  199
#> 201      35831461  871176.816  2346467.156 2.795089e+07            0  200