Model an SEIR-V epidemic with interventions

Simulate an epidemic using a deterministic, compartmental epidemic model with the compartments "susceptible", "exposed", "infectious", "recovered", and "vaccinated". This model can accommodate heterogeneity in social contacts among demographic groups, as well as differences in the sizes of demographic groups.

The population and an infection arguments are mandatory, while passing an intervention and vaccination are optional and can be used to simulate scenarios with different epidemic responses or different levels of the same type of response. See Details for more information.

Usage

epidemic_default_cpp(
  population,
  infection,
  intervention = NULL,
  vaccination = NULL,
  time_end = 100,
  increment = 1
)

epidemic_default_r(
  population,
  infection,
  intervention = NULL,
  vaccination = NULL,
  time_end = 100,
  increment = 1
)

Arguments

population

An object of the population class, which holds a population contact matrix, a demography vector, and the initial conditions of each demographic group. See population().

infection

An infection object created using infection(). Must have the basic reproductive number \(R_0\) of the infection, the infectious period, and the pre-infectious period. These are used to calculate the transmission rate \(\beta\), the rate at which individuals move from the 'exposed' to the 'infectious' compartment, \(\alpha\), and the recovery rate \(\gamma\).

intervention

An <intervention> object representing an optional non-pharmaceutical intervention applied to the population during the epidemic. See intervention() for details on constructing interventions with age-specific effects on social contacts, as well as for guidance on how to concatenate multiple overlapping interventions into a single <intervention> object.

vaccination

A <vaccination> object representing an optional vaccination regime with a single dose, followed during the course of the epidemic, with a start and end time, and age-specific vaccination rates. See vaccination().

time_end

The maximum number of timesteps over which to run the model. Taken as days, with a default value of 200 days.

increment

The size of the time increment. Taken as days, with a default value of 1 day.

Value

A data.table with the columns "time", "compartment", "age_group", "value". The compartments correspond to the compartments of the model chosen with model. The current default model has the compartments "susceptible", "exposed", "infectious", "recovered", and "vaccinated".

Details

epidemic_default_cpp() is a wrapper function for .epidemic_default_cpp(), an internal C++ function that uses Boost odeint solvers for an SEIR-V model . .epidemic_default_cpp() accepts arguments that are created by processing the population, infection, intervention and vaccination arguments to the wrapper function into simpler forms.

epidemic_default_r() is a wrapper around the internal function .ode_epidemic_default(), which is passed to deSolve::lsoda().

Both models return equivalent results, but the C++ implementation is faster.

This model only allows for single, population-wide rates of transition between the 'susceptible' and 'exposed' compartments, between the 'exposed' and 'infectious' compartments, and in the recovery rate.