Model an Ebola virus disease epidemic

Simulate an epidemic using a discrete-time, stochastic SEIR compartmental model with compartments based on Li et al. (2019), and with Erlang passage times based on a model developed by Getz and Dougherty (2017), developed to model the West African Ebola virus disease outbreak of 2014. See Details for more information.

epidemic_ebola_cpp() is an Rcpp implementation of this model that currently lags behind the R implementation, and is likely to be removed.

Usage

epidemic_ebola_r(population, infection, intervention = NULL, time_end = 100)

epidemic_ebola_cpp(population, infection, time_end = 100)

Arguments

population

An object of the <population> class, see population().

This model only accepts a <population> without demographic structure, that is, the demography_vector must be a single number representing the total size of the affected population.

The model also does not account for demographic differences in social contacts, which means that the contact_matrix is ignored. For consistency, the matrix must be square and have as many rows as demography groups, which is one.

infection

An <infection> object created using infection(). Must have:

r0, a single number for the basic reproductive number \(R_0\) of the infection,

infectious_period, a single number for the infectious period, taken to be in days,

preinfectious_period, a single number for the pre-infectious period before the onset of symptoms, taken to be in days,

prop_hospitalised, a single number in the range 0.0 -- 1.0 for the proportion of infectious (assumed symptomatic) individuals who are hospitalised,

etu_safety, a single number in the range 0.0 -- 1.0 for the relative effectiveness of hospitalisation (in an Ebola Treatment Unit; ETU) in reducing transmission between hospitalised individuals and susceptible ones. 0.0 would indicate that hospitalisation does not reduce transmission, while 1.0 would indicate that it completely prevents transmission.

funeral_safety, a single number in the range 0.0 -- 1.0 for the relative efficacy of practices designed to prevent transmission of Ebola virus disease associated with funerals; alternatively interpretable as the proportion of funerals following these practices.

r0, infectious_period, and preinfectious_period are used to calculate the baseline transmission rate \(\beta\), as well as the rates \(\gamma^E\) and \(\gamma^I\) at which individuals move from the 'exposed' to the 'infectious' compartment, and from the 'infectious' to the 'recovered' compartment, respectively. See Details for more information.

intervention

An optional <rate_intervention> object representing pharmaceutical or non-pharmaceutical interventions applied to the infection's parameters, such as the transmission rate, over the epidemic. See intervention() for details on constructing rate interventions.

Currently, only interventions on the transmission rate beta are supported, and should be passed as list(beta = intervention(type = "rate", ...)). Defaults to NULL, representing no interventions on model parameters.

time_end

The maximum number of timesteps over which to run the model, in days. Defaults to 100 days.

Value

A <data.table> in long format with the columns "time", "compartment", "age_group", and "value", that gives the number of individuals in each model compartment over time ( from 1 to time_end).

Details: Discrete-time ebola virus disease model

This model has compartments adopted from the consensus model for Ebola virus disease presented in Li et al. (2019), and with transitions between epidemiological compartments modelled using Erlang sub-compartments adapted from Getz and Dougherty (2018); see References.

The R code for this model is adapted from code by Hạ Minh Lâm and initially made available on Epirecipes (https://github.com/epirecipes/epicookbook) under the MIT licence.

The model implementation differs from Getz and Dougherty's (2018) in allowing users to set the basic reproductive number \(R_0\), and the mean infectious (\(\rho^I\) in Getz and Dougherty) and pre-infectious periods in days (\(\rho^E\) in Getz and Dougherty). Getz and Dougherty instead calculate these from other model parameters.

The shape of the Erlang distributions of passage times through the exposed and infectious compartments (\(k^E\) and \(k^I\)) are fixed to 2 (this was allowed to vary in Getz and Dougherty).

The transition rates between the exposed and infectious, and infectious and funeral compartments, \(\gamma^E\) and \(\gamma^I\) in Getz and Dougherty's notation, are calculated following their equation (6). $$\gamma^E = \dfrac{k^E}{\rho^E} = \dfrac{2}{\rho^E} ~\text{and}~ \gamma^I = \dfrac{k^I}{\rho^I} = \dfrac{2}{\rho^I}$$

In this discrete time model, \(\gamma^E\) and \(\gamma^I\) are used to determine the number of Erlang sub-compartments in each epidemiological compartment, and the probability of newly exposed or infectious individuals beginning in one of the compartments (thus allowing for variation in passage times).

Hospitalisation, funerals, and removal

Infectious individuals have a probability of prop_hospitalised of being transferred to the hospitalised compartment. This compartment has the same number of sub-compartments, which means that an infectious individual with \(N\) timesteps before exiting the infectious compartment will exit the hospitalised compartment in the same time.

Hospitalised individuals can contribute to transmission of Ebola to susceptibles depending on the value of etu_safety passed as part of the infection argument, which scales the baseline transmission rate \(\beta\) for hospitalised individuals.

We assume that deaths in hospital lead to Ebola-safe funerals, and individuals exiting the hospitalised compartment move to the 'removed' compartment, which holds both recoveries and deaths.

We assume that deaths outside of hospital lead to funerals that are potentially not Ebola-safe, and the funeral_safety passed as part of the infection argument scales the baseline transmission rate \(\beta\) for funeral transmission of Ebola to susceptibles.

Individuals are assumed to spend only a single timestep in the funeral transmission compartment, before they move into the 'removed' compartment.

Individuals in the 'removed' compartment do no affect model dynamics.

References

Li, S.-L., Ferrari, M. J., Bjørnstad, O. N., Runge, M. C., Fonnesbeck, C. J., Tildesley, M. J., Pannell, D., & Shea, K. (2019). Concurrent assessment of epidemiological and operational uncertainties for optimal outbreak control: Ebola as a case study. Proceedings of the Royal Society B: Biological Sciences, 286(1905), 20190774. doi:10.1098/rspb.2019.0774

Getz, W. M., & Dougherty, E. R. (2018). Discrete stochastic analogs of Erlang epidemic models. Journal of Biological Dynamics, 12(1), 16–38. doi:10.1080/17513758.2017.1401677