Get the epidemic size

Get the size of the epidemic at any stage between the start and the end. This is calculated as the number of individuals recovered from infection at that stage of the epidemic.

Usage

epidemic_size(data, stage = 1, by_group = TRUE, deaths = TRUE)

Arguments

data

A data.table (or data.frame) of model output, typically the output of epidemic_default_cpp() or similar functions.

stage

The stage of the epidemic at which to return the epidemic size; here, 0.0 represents the initial conditions of the epidemic (0% of model time ), while 1.0 represents the end of the epidemic model (100% of model time). The values returned at stage = 1.0 represent the final size of the epidemic.

by_group

A logical representing whether the epidemic size should be returned by demographic group, or whether a single population-wide value is returned.

deaths

A logical value that indicates whether to count individuals in the epidemic size calculation. Setting deaths = TRUE looks for a "dead" compartment in the data. If there is no such column, the function returns only the final number of recovered individuals in each demographic group.

Value

A single number when by_group = FALSE, or a vector of numbers of the same length as the number of demographic groups when by_group = TRUE.

Details

This function can be used to calculate the final size of the epidemic, by setting stage = 1.0 (100% of model time; the default).

The function allows for the calculation of epidemic sizes by demographic group as well as the total epidemic size.

Examples

# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# create an infection object
pandemic_influenza <- infection(
  r0 = 1.5, infectious_period = 7, preinfectious_period = 3
)

# run epidemic simulation with no vaccination or intervention
data <- epidemic_default_cpp(
  population = uk_population,
  infection = pandemic_influenza,
  time_end = 200,
  increment = 1
)

# get the final epidemic size
epidemic_size(data)
#> No 'dead' compartment found in `data`; counting only 'recovered' individuals in the epidemic size.
#> demo_group_1 
#>     27950895 

# get the epidemic size at the halfway point
epidemic_size(data, stage = 0.5)
#> No 'dead' compartment found in `data`; counting only 'recovered' individuals in the epidemic size.
#> demo_group_1 
#>      1132229