epidemics: Composable Epidemic Scenario Modelling

A library of compartmental epidemic models taken from the published literature, and classes to represent affected populations, public health response measures including non-pharmaceutical interventions on social contacts, non-pharmaceutical and pharmaceutical interventions that affect disease transmissibility, vaccination regimes, and disease seasonality, which can be combined to compose epidemic scenario models.

See also

Author

Maintainer: Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID) [copyright holder]

Authors:

Rosalind Eggo rosalind.eggo@lshtm.ac.uk (ORCID) [copyright holder]

Edwin Van Leeuwen edwin.vanleeuwen@ukhsa.gov.uk (ORCID) [copyright holder]

Other contributors:

Adam Kucharski adam.kucharski@lshtm.ac.uk (ORCID) [contributor, reviewer]

Tim Taylor tim.taylor@hiddenelephants.co.uk (ORCID) [contributor, reviewer]

Banky Ahadzie bahadzie@gmail.com [contributor]

Hugo Gruson hugo.gruson@data.org (ORCID) [reviewer]

Joshua W. Lambert joshua.lambert@lshtm.ac.uk (ORCID) [reviewer]

James M. Azam james.azam@lshtm.ac.uk (ORCID) [reviewer]