Prepare an infection for an epidemic model

Prepare an S3 object of the <infection> class, which is used to hold some basic parameters of the infection that is to be modelled as causing an epidemic.

<infection>s can hold any number of extra parameters as required by specific epidemic models.

Usage

infection(name = NA_character_, r0, infectious_period, ...)

is_infection(x)

Arguments

name

String for the name of the infection. The default value is NA.

r0

The basic reproductive number of the infection, \(R_0\). Must be a number or a vector of numbers of the same length as the number of demographic groups, depending on the model being implemented. The 'default' model requires a single number.

infectious_period

The mean infections period. Must be a number or a vector of numbers of the same length as the number of demographic groups, depending on the model being implemented. Currently, only the option to pass a single infectious period common to all demographic groups is available.

...

Other arguments representing other infection parameters. This argument is intentionally left flexible to allow for single values or vectors of values.

x

An object to be checked as being an <infection>.

Value

infection() returns an object of the <infection> S3 class.

is_infection() returns a logical for whether the object is of the <infection> class.

Examples

pandemic_influenza <- infection(
  name = "pandemic influenza",
  r0 = 1.3,
  infectious_period = 5,
  preinfectious_period = 2
)

# print to examine
pandemic_influenza
#> <infection> object
#> 
#>  Name: 
#> pandemic influenza
#> 
#>  R0: 
#> 1.3
#> 
#>  Infectious period: 
#> 5
#> 
#>  Other parameters: 
#> "preinfectious_period"

# check for the infection class
is_infection(pandemic_influenza)
#> [1] TRUE