Get new infections

Get new infections

Usage

new_infections(data, compartments_from_susceptible, by_group = TRUE)

Arguments

data

A data.table (or data.frame) of model output, typically the output of epidemic_default_cpp() or similar functions.

compartments_from_susceptible

An optional argument, for a character vector of the names of model compartments into which individuals transition from the "susceptible" compartment, and which are not related to infection. A common example is a compartment for "vaccinated" individuals who are no longer susceptible, but who should also not be counted as infected.

by_group

A logical representing whether the epidemic size should be returned by demographic group, or whether a single population-wide value is returned.

Value

A data.table with the same columns as data, but with the additional variable under compartment, "new_infections", resulting in additional rows.

Examples

# create a population
uk_population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0, 0),
    nrow = 1, ncol = 5L
  )
)

# create an infection
pandemic_influenza <- infection(
  r0 = 1.5, infectious_period = 7, preinfectious_period = 3
)

# run epidemic simulation with no vaccination or intervention
data <- epidemic_default_cpp(
  population = uk_population,
  infection = pandemic_influenza,
  time_end = 200,
  increment = 1
)

new_infections(data)
#> Loading required namespace: testthat
#>      time demography_group new_infections
#>   1:    0     demo_group_1         0.0000
#>   2:    1     demo_group_1       206.2792
#>   3:    2     demo_group_1       513.2870
#>   4:    3     demo_group_1       713.8049
#>   5:    4     demo_group_1       851.9442
#>  ---                                     
#> 197:  196     demo_group_1    301353.1136
#> 198:  197     demo_group_1    294124.8571
#> 199:  198     demo_group_1    286895.5505
#> 200:  199     demo_group_1    279678.9433
#> 201:  200     demo_group_1    272487.9871