SIR epidemic model using 'deSolve'

SIR epidemic model using 'deSolve'

Usage

sir_desolve(
  times = seq(0, 200, length.out = 2001),
  init = initial_conditions(p_initially_infected = 0.01),
  parms = c(beta = 0.1, gamma = 0.05)
)

Arguments

times

A sequence of timepoints at which to solve for the proportion of the population in each state (susceptible, infected, or recovered).

init

The initial conditions of the population: a named vector with the proportions of the population in each state ("S": susceptible, "I": infected, or "R": recovered). Can be easily constructed using the helper function [initial_conditions()].

parms

The infection parameters in the form of a named vector with an element "beta" for the rate of transmission \(\beta\), and "gamma" for the rate of recovery of infected individuals \(\gamma\). Formally, both \(beta\) and \(gamma\) are the rate parameters of separate exponential distributions.

Value

A data table with the proportion of the population in the susceptible, infected, and recovered states at each time point in times. The data are returned in tidy format, and have \(3T\) rows, where \(T\)

is the number of timepoints.

Examples

library(epidemics)
sir_desolve(
  times = seq(0, 100, length.out = 100),
  init = initial_conditions(p_initially_infected = 0.01),
  parms = c(beta = 0.5, gamma = 0.05)
)
#>            time state proportion
#>   1:   0.000000     S  0.9900000
#>   2:   1.010101     S  0.9837094
#>   3:   2.020202     S  0.9739625
#>   4:   3.030303     S  0.9590074
#>   5:   4.040404     S  0.9363926
#>  ---                            
#> 296:  95.959596     R  0.9857767
#> 297:  96.969697     R  0.9864747
#> 298:  97.979798     R  0.9871383
#> 299:  98.989899     R  0.9877692
#> 300: 100.000000     R  0.9883691