Model an SIRV epidemic

Function for a deterministic susceptible-infectious-recovered model with an optional vaccination component. Allows heterogeneity in social contact patterns, the demography distribution, and key epidemiological parameters: the reproductive number \(R_0\), and the infectious period \(1/\gamma\). Also allows for group-specific initial proportions in each model compartment, as well as group-specific vaccination start dates and vaccination rates.

Usage

epi_demic(
  population,
  R0 = 1.5,
  infectious_period = 7,
  intervention,
  nu = 0.01,
  t_vax_begin = t_max * 0.6,
  t_vax_end = t_max * 0.8,
  t_max = 200,
  t_increment = 1
)

Arguments

population

An object of the population class, which holds a population contact matrix, a demography vector, and the initial conditions of each demographic group. See population().

R0

The reproductive number of the infection. May be a single number or a vector of the same length as the number of demographic groups.

infectious_period

The mean infections period. May be a single number or a vector of the same length as the number of demographic groups.

intervention

An object of the intervention class that specifies the start and end of the intervention, and the effect of the intervention on each demographic group. See intervention().

nu

The vaccination rate \(\nu\). May be a single number or a vector of the same length as the number of demographic groups.

t_vax_begin

The time at which vaccination begins. Must be a number, less than t_max. Defaults to 60% of the maximum time.

t_vax_end

The time at which vaccination ends. Defaults to 80% of the maximum time.

t_max

The maximum number of timesteps over which to run the model.

t_increment

The size of the time increment.

Value

A data.frame with the columns "time", "compartment", "age_group", "value". The comparments ar "susceptible", "infected", "recovered", and "vaccinated".

Examples

# create a population
population <- population(
  name = "UK population",
  contact_matrix = matrix(1),
  demography_vector = 67e6,
  initial_conditions = matrix(
    c(0.9999, 0.0001, 0, 0),
    nrow = 1, ncol = 4
  )
)

# set up intervention
new_intervention <- intervention(
  name = "close schools",
  time_begin = 50,
  time_end = 80,
  contact_reduction = c(0.2)
)

# run epidemic simulation
epi_demic(
  population = population,
  R0 = 1.5, infectious_period = 7,
  intervention = new_intervention,
  nu = 0.01,
  t_vax_begin = 100,
  t_vax_end = 200,
  t_max = 200,
  t_increment = 1
)
#>      time compartment age_group    value
#>   1:    0 susceptible         1 66993300
#>   2:    1 susceptible         1 66991812
#>   3:    2 susceptible         1 66990214
#>   4:    3 susceptible         1 66988497
#>   5:    4 susceptible         1 66986654
#>  ---                                    
#> 800:  196  vaccinated         1 30242404
#> 801:  197  vaccinated         1 30405281
#> 802:  198  vaccinated         1 30566523
#> 803:  199  vaccinated         1 30726150
#> 804:  200  vaccinated         1 30884178