Construct a new intervention for an epidemic model

Prepare an <intervention> object that specifies a non-pharmaceutical intervention regime that reduces contacts, for use in an epidemic model. Interventions have a single start and end time that applies to all demographic groups in the population, but can have groups-specific effects on the reduction of contacts.

Combine <intervention> objects to create sequential and overlapping intervention regimes using c() on two or more <intervention> objects.

Usage

intervention(name = NA_character_, time_begin, time_end, contact_reduction)

is_intervention(x)

no_intervention(population)

# S3 method for intervention
c(x, ...)

Arguments

name

String for the name of the intervention.

time_begin

Single number for the start time of the intervention.

time_end

Single number for the end time of the intervention.

contact_reduction

A matrix \([i, 1]\) with as many rows as the number of demographic groups in the target population, and a single column. Each element gives the group-specific proportion reduction in contacts.

When multiple interventions are combined using c(), they are stacked column wise to form a matrix \([i, j]\). See Details for how the effect of intervention \(j\) on group \(i\) is handled in epidemic models.

x

An <intervention> object, or an object to be checked as an <intervention> object.

population

A <population> object with a contact_matrix member.

...

intervention objects to combine with x to create a multi-dose <intervention> object.

Value

An object of the <intervention> S3 class.

Concatenating two or more <intervention> objects using c() also returns a <intervention> object. This object holds the intervention-specific start and end times, and group-specific contact reductions specified by all the constituent intervention actions. The combined effect of these actions on the population is handled internally by epidemic model functions.

A "null" intervention generated using no_intervention(population) returns an <intervention> that has start and end times, and contact reduction all set to 0.0.

is_intervention() returns a logical for whether the object is of the <intervention> class.

Details

Epidemic models that can accommodate interventions are able to accommodate any number of interventions with different start and end times and different group-specific effects.

Models such as epidemic_default_cpp() are set up to treat interventions with overlapping periods (i.e., overlap between the time when they are active ) as having an additive effect on contact reductions. The group-specific effect of \(J\) overlapping interventions is thus a vector \(\sum_{j = 1}^J x_{ij}\). This is handled internally by the epidemic model code.

For example, a contact reduction matrix for two perfectly overlapping interventions (\(J = 2\)) with different effects across three demographic groups (\(I = 3\)) would be represented as:

\(\begin{bmatrix}0.1 & 0.05\\0.1 & 0.1\\0.1 & 0.0\end{bmatrix}\)

In epidemic models, the cumulative group-specific effect when both interventions are active would be \((0.15, 0.2, 0.1)\).

Examples

# assuming a population with two age groups, 0 -- 18, and 18+
# an example in which schools are closed for 30 days (or other time units)
close_schools <- intervention(
  name = "close schools",
  time_begin = 50,
  time_end = 80,
  contact_reduction = matrix(c(0.5, 0.01)) # reduces contacts differentially
)
close_schools
#> <intervention>
#> Intervention name: "close schools"
#> Time begin:
#>      [,1]
#> [1,]   50
#> Time end:
#>      [,1]
#> [1,]   80
#> Contact reduction:
#>      [,1]
#> [1,] 0.50
#> [2,] 0.01

# Check for intervention class
is_intervention(close_schools)
#> [1] TRUE

# Concatenating interventions
# create first intervention
npi_1 <- intervention(
  time_begin = 30,
  time_end = 60,
  contact_reduction = matrix(0.1)
)

# second intervention
npi_2 <- intervention(
  time_begin = 45,
  time_end = 75,
  contact_reduction = matrix(0.1)
)

c(npi_1, npi_2)
#> <intervention>
#> Intervention name: NA
#> Time begin:
#>      npi_1 npi_2
#> [1,]    30    45
#> Time end:
#>      npi_1 npi_2
#> [1,]    60    75
#> Contact reduction:
#>      npi_1 npi_2
#> [1,]   0.1   0.1

# A null intervention for scenarios without interventions
pop <- population(
  demography_vector = 1e6, contact_matrix = matrix(1),
  initial_conditions = matrix(c(0.99, 0.01, 0.0), nrow = 1)
)

no_intervention(pop)
#> Intervention: some `time_end`s are not greater than `time_begin`s
#> <intervention>
#> Intervention name: "no_intervention"
#> Time begin:
#>      [,1]
#> [1,]    0
#> Time end:
#>      [,1]
#> [1,]    0
#> Contact reduction:
#>      [,1]
#> [1,]    0
close_schools <- intervention(
  name = "close schools",
  time_begin = 50,
  time_end = 80,
  contact_reduction = matrix(c(0.5, 0.01)) # reduces contacts differentially
)
is_intervention(close_schools)
#> [1] TRUE